Linter: Extract `fixabilityFor` out of the CLI and export it
`fixabilityFor` is the single place that decides whether an offense is safely autocorrectable, unsafely autocorrectable, or neither. It lived under `src/cli/` but has no CLI-specific behavior, so move it to the package root and export it for other consumers.
