+++ javascript/packages/linter/src/cli/file-processor.ts
@@ -11,13 +11,13 @@ import { fileURLToPath } from "node:url"
 import { availableParallelism } from "node:os"
 import { colorize } from "@herb-tools/highlighter"
 import { deserializeDiagnostic, didyoumean } from "@herb-tools/core"
-import { fixabilityFor } from "./fixability.js"
+import { fixabilityFor } from "../fixability.js"
 
 import type { Diagnostic } from "@herb-tools/core"
 import type { FormatOption } from "./argument-parser.js"
 import type { HerbConfigOptions } from "@herb-tools/config"
 import type { WorkerInput, WorkerResult } from "./lint-worker.js"
-import type { Fixability } from "./fixability.js"
+import type { Fixability } from "../fixability.js"
 import type { VersionSkippedRule } from "../linter.js"
 import type { LintOffense, RuleClass } from "../types.js"
